Refactored some player endpoints
BUT WAIT THERE'S MORE! APPLY TODAY FOR YOUR CHANCE TO BE ENTERED INTO
OUR Â5000 SWEEPSTAKES! That's actually a shitty deal. You're just
applying for the chance to get the chance to win the sweepstakes. smh my
head... But yeah the endpoint exists to get all of the connected
players. Plus some player endpoints are no longer admin endpoints,
because it turns out clients are greedy shits and want them too. Sick...
(that is a peak 2017 Victor commit if I've ever seen one :D)